+++ src/BusinessLogic/AbstractClassAction.php
@@ -10,7 +10,12 @@ abstract class AbstractClassAction
 
     abstract public static function getInteriorName(): string;
 
-    abstract public static function roleCheck(int $user_id, string $task_status, bool $availability = null): bool;
-//    $availability - не для всех методов(только для действия ОТКЛИКНУТЬСЯ) - это доступность
-//    для исполнителя откликаться на задание, ёе может не быть, потому что заказчик уже 1 раз отказал ему
+    public static function getClassName(): string
+    {
+        return static::class;
+    }
+
+    abstract public static function roleCheck(int $user_id, int $customer_id, int $worker_id): bool;
 }
+
+//string $task_status, bool $availability = null
