+++ src/BusinessLogic/RefuseAction.php
@@ -19,12 +19,10 @@ class RefuseAction extends AbstractClassAction
         return 'act_refuse';
     }
 
-    public static function roleCheck(int $user_id, string $task_status, bool $availability = null): bool
+    public static function roleCheck(int $user_id, int $customer_id, int $worker_id): bool
     {
-        if ($task_status == 'IN_WORKING') {
-            if ($user_id == 'worker_id') {
-                return TRUE;
-            }
+        if ($user_id == $worker_id) {
+            return TRUE;
         }
         return NULL;
     }
